Calculates percent of population using internet
This section of the code calculates the percentage of a country's
population that uses the internet, using the values from the origional
number of people using the internet and population text files. It
appends this new value into the existing list. The values are truncated
to two decimal places right now, future versions will round instead.